+++ crates/teloxide-core/CHANGELOG.md
@@ -43,8 +43,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add `BusinessConnection`, `BusinessMessage`, `EditedBusinessMessage` and `DeletedBusinessMessages` variants to `UpdateKind` enum
 
 - `ApiError::BotKickedFromChannel` ([#1157][pr1157])
+- `From<&Message> for MessageId` impl ([#1271][pr1271])
 
 [pr1157]: https://github.com/teloxide/teloxide/pull/1157
+[pr1271]: https://github.com/teloxide/teloxide/pull/1271
 
 ### Changed